backend/requirements.txt
pandas
numpy
xgboost
scikit-learn
tensorflow
fastapi
pydantic
uvicorn
python-dotenv
joblib
google-generativeai